apps/api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
 "fastapi[standard]>=0.120.0",
 "langchain>=1.0.2",
 "langchain-google-genai>=3.0.0",
 "nylas>=6.13.1",
 "python-dateutil>=2.8.2",
 "python-dotenv>=1.1.1",
 "supabase>=2.22.2",
 "httpx>=0.25.0",
]
